+++ cas-20200630/alibabacloud_cas20200630/models/_list_cert_response_body.py
@@ -111,7 +111,7 @@ class ListCertResponseBodyList(DaraModel):
         common_name: str = None,
         custom_identifier: str = None,
         extra: str = None,
-        id: str = None,
+        id: int = None,
         identifier: str = None,
         key_exportable: bool = None,
         organization: str = None,